First thing here is that it updates scripts/test1.sh so that once again it reports speed ratios between the systems tested. If prefixes the ratio with a "?" if the test time is under 250ms because for brief tests the time measurements may be unreliable. I have bene working on the conservative CSL branch - and in this version newcslgc.cpp will not compile. I will fix this soon! I re-wrote conservative.txt to reflect my current thoughts and have been trying to rationalise the names of some fields within structures and some external variables. And while I am at it I continue migrating the code style from C to what I take to be C++ which includes trying to remove use of "#define" where that is not too uncomfortable. Doing so in favour of use of constexpr names allows the compiler to do a bit more type-checking so there are places where in consequence I made a few things size_t not int so they were unsigned. I also provide a new profile.dat. That will typically get picked up if you delete the cslbuild directory and start again with ./configure. But if that is too radical just copy it from csl/generated-c into cslbuild/generated-c. It should get CSL compilation into C++ a little more up to date with changes in the Reduce source files.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5781 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/cslgc.cpp
@@ -501,7 +501,7 @@ static void real_garbage_collector()
 {
 // I lift the real garbage collector to a separate function mainly
 // so that I can set breakpoints on it!
-    for (int i=0; i<=LOG2_VECTOR_CHUNK_BYTES; i++)
+    for (size_t i=0; i<=LOG2_VECTOR_CHUNK_BYTES; i++)
         free_vectors[i] = 0;
 
     cons_cells = symbol_heads = strings = user_vectors =
